+++ HermesProxy/World/Server/Packets/MailCreateTextItem.cs
@@ -14,6 +14,6 @@ public class MailCreateTextItem : ClientPacket
 	public override void Read()
 	{
 		this.Mailbox = base._worldPacket.ReadPackedGuid128();
-		this.MailID = base._worldPacket.ReadUInt32();
+		this.MailID = (uint)base._worldPacket.ReadUInt64();
 	}
 }
