+++ HermesProxy/World/Server/Packets/MailListResult.cs
@@ -16,7 +16,7 @@ public class MailListResult : ServerPacket
 
 	public override void Write()
 	{
-		base._worldPacket.WriteInt32(this.Mails.Count);
+		base._worldPacket.WriteUInt32((uint)this.Mails.Count);
 		base._worldPacket.WriteInt32(this.TotalNumRecords);
 		this.Mails.ForEach(delegate(MailListEntry p)
 		{
