+++ HermesProxy/World/Server/Packets/MailTakeItem.cs
@@ -16,7 +16,7 @@ public class MailTakeItem : ClientPacket
 	public override void Read()
 	{
 		this.Mailbox = base._worldPacket.ReadPackedGuid128();
-		this.MailID = base._worldPacket.ReadUInt32();
-		this.AttachID = base._worldPacket.ReadUInt32();
+		this.MailID = (uint)base._worldPacket.ReadUInt64();
+		this.AttachID = (uint)base._worldPacket.ReadUInt64();
 	}
 }
